move the tabs of training center's course into  trainingCenter/course folder, create import files fix imports
