+++ frontend/src/pages/trainingCenter/course/aboutCourse/AboutCourse.tsx
