+++ frontend/src/pages/trainingCenter/course/aboutCourse/aboutCourse.module.scss
