+++ frontend/src/pages/trainingCenter/course/aboutCourse/index.ts
