+++ frontend/src/pages/trainingCenter/course/assignments-grades/assignments-grades.module.scss
