+++ frontend/src/pages/trainingCenter/course/assignments-grades/index.ts
