+++ frontend/src/pages/trainingCenter/course/index.ts
@@ -1,0 +1,5 @@
+export * from './aboutCourse'
+export * from './assignments-grades'
+export * from './program'
+export * from './studentsList'
+export * from './TrainingCenterCourse'
