+++ frontend/src/pages/trainingCenter/course/program/index.ts
@@ -1,0 +1,4 @@
+export * from './Program'
+export * from './block-header/BlockHeader'
+export * from './expanded-content/ExpandedContent'
+export * from './lesson-content/LessonContent'
