+++ frontend/src/pages/trainingCenter/course/studentsList/studentBlock/index.ts
