+++ frontend/src/pages/trainingCenter/course/studentsList/studentBlock/studentBlock.module.scss
